requirements.txt
# for development
# pip
# flake8
# black
# isort
# natsort
# pandas
# matplotlib
# jupyterlab
# scikit-learn
# torchinfo
# labelImg

# for deployment
numpy
streamlit==1.19.0
streamlit-webrtc==0.45.0
opencv-python-headless
Pillow
torch
torchvision
